% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_manage.R
\name{app_create}
\alias{app_create}
\title{Functions to manage shiny applications}
\usage{
app_create(app_name, project_id, api_key, host)
}
\arguments{
\item{api_key}{Domino Data Labs user api key}

\item{host}{Domino Data Labs host}

\item{project_name}{name of parent project}

\item{owner_name}{name of owner}
}
\description{
Functions to manage shiny applications
}
\author{
Michael B. Garcia
}
