% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_manage.R
\name{app_get_url}
\alias{app_get_url}
\title{Functions to manage shiny applications}
\usage{
app_get_url(app_id, host)
}
\arguments{
\item{host}{Domino Data Labs host}

\item{project_name}{name of parent project}

\item{owner_name}{name of owner}

\item{api_key}{Domino Data Labs user api key}
}
\description{
Functions to manage shiny applications
}
\author{
Michael B. Garcia
}
