% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_init.R
\name{app_init}
\alias{app_init}
\title{Create application instance}
\usage{
app_init(project_name, project_id, owner_name, owner_id, hardware_tier_id,
  hrs_expire, api_key, host)
}
\arguments{
\item{project_name}{name of parent project}

\item{project_id}{id of parent project}

\item{owner_name}{name of project owner}

\item{owner_id}{id of project owner}

\item{hardware_tier_id}{hardware tier selected for application}

\item{hrs_expire}{hours to keep application open}

\item{api_key}{Domino Data Labs user api key}

\item{host}{Domino Data Labs host}
}
\description{
Create application instance
}
\author{
Michael B. Garcia
}
