% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{ep_self}
\alias{ep_self}
\title{Domino Data Labs API Endpoints}
\usage{
ep_self()
}
\arguments{
\item{project_id}{project id of a Domino project}

\item{app_id}{app id of a published Shiny app}

\item{project_name}{name of parent project}

\item{owner_name}{name of owner}

\item{file_path}{destination path used for file upload}
}
\value{
url endpoints of Domino Data Labs API
}
\description{
These endpoints are used by functions within this package
}
\author{
Michael B. Garcia
}
