% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_manage.R
\name{projects_get}
\alias{projects_get}
\title{Functions to manage Domino projects}
\usage{
projects_get(api_key, host)
}
\arguments{
\item{api_key}{Domino Data Labs user api key}

\item{host}{Domino Data Labs host}

\item{project_name}{name of project}

\item{project_id}{id of project}

\item{owner_name}{name of owner}

\item{owner_id}{id of owner}

\item{hrs_expire}{hours to keep application open}

\item{hash}{hashed id to be used for project name}
}
\description{
Functions to manage Domino projects
}
\author{
Michael B. Garcia
}
