% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh_main_tbl.R
\name{refresh_main_tbl}
\alias{refresh_main_tbl}
\title{Refresh project table}
\usage{
refresh_main_tbl(project_name, api_key, host)
}
\arguments{
\item{project_name}{name of parent project}

\item{api_key}{Domino Data Labs user api key}

\item{host}{Domino Data Labs host}

\item{project_id}{id of parent project}

\item{owner_name}{name of project owner}

\item{owner_id}{id of project owner}

\item{hardware_tier_id}{hardware tier selected for application}

\item{hrs_expire}{hours to keep application open}
}
\value{
success or failure messages
}
\description{
Refreshes tbl in the master app control center
}
\author{
Michael B. Garcia
}
