% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinydominoes_init.R
\name{shinydominoes_init}
\alias{shinydominoes_init}
\title{Initiate shinydominoes on an existing Shiny application}
\usage{
shinydominoes_init(shiny_app)
}
\arguments{
\item{shiny_app}{path to the shiny app to deploy}
}
\value{
success or failure messages
}
\description{
This function sets up the necessary dependency files needed
in your project to scale out your Shiny application to multiple
project nodes
}
\author{
Michael B. Garcia
}
