% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ep_self}
\alias{ep_self}
\title{Functions}
\usage{
ep_self()
}
\arguments{
\item{project_name}{name of parent project}

\item{owner_name}{name of owner}

\item{api_key}{Domino Data Labs user api key}

\item{host}{Domino Data Labs host}
}
\value{
confirmation message of success or failure
}
\description{
These functions will be broken out
}
\author{
Michael B. Garcia
}
